+++ src/main/java/roomescape/web/page/AdminController.java
@@ -1,4 +1,4 @@
-package roomescape.controller.page;
+package roomescape.web.page;
 
 import org.springframework.stereotype.Controller;
 import org.springframework.web.bind.annotation.GetMapping;
