+++ src/main/java/roomescape/web/request/ReservationRequest.java
@@ -1,4 +1,4 @@
-package roomescape.controller.request;
+package roomescape.web.request;
 
 import roomescape.domain.Reservation;
 import roomescape.domain.ReservationTime;
