+++ src/main/java/roomescape/web/request/ReservationTimeRequest.java
@@ -1,4 +1,4 @@
-package roomescape.controller.request;
+package roomescape.web.request;
 
 import roomescape.domain.ReservationTime;
