+++ src/main/java/roomescape/web/request/ThemeRequest.java
@@ -1,4 +1,4 @@
-package roomescape.controller.request;
+package roomescape.web.request;
 
 import java.util.Objects;
 import roomescape.domain.Theme;
